Add: Account test, add: Checking account, add: Savings Bonds
Create account test that contains the main method.
Create checking account class to make two validations, one if the balance is too low cant make a withdraw and the second one if the user enter a negative value, the method will print the balance, the valur of the withdraw and the new balance.
Create savings bond class that make a validations depends on the term of the saving bond later create a new method that shows the balance plus interest rate the interest rate and the months remaining.